# public/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Widget": {
        "properties": {
          "config": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "dashboardId": {
            "format": "uuid",
            "type": "string"
          },
          "height": {
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "positionX": {
            "type": "integer"
          },
          "positionY": {
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "widgetType": {
            "type": "string"
          },
          "width": {
            "type": "integer"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "description": "OpenAPI specification covering the available API routes for Forge.",
    "title": "Forge API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/widgets": {
      "delete": {
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            },
            "description": "Widget deleted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing widget id"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Widget not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Delete a widget"
      },
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dashboardId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Widget"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Matching widgets"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing query parameter"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "List widgets for a user or dashboard"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "config": {
                    "type": "object"
                  },
                  "dashboardId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "height": {
                    "type": "integer"
                  },
                  "positionX": {
                    "type": "integer"
                  },
                  "positionY": {
                    "type": "integer"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "widgetType": {
                    "type": "string"
                  },
                  "width": {
                    "type": "integer"
                  }
                },
                "required": [
                  "userId",
                  "dashboardId",
                  "widgetType",
                  "height",
                  "width",
                  "positionX",
                  "positionY"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            },
            "description": "Widget created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Create a widget"
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "config": {
                    "type": "object"
                  },
                  "height": {
                    "type": "integer"
                  },
                  "id": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "positionX": {
                    "type": "integer"
                  },
                  "positionY": {
                    "type": "integer"
                  },
                  "width": {
                    "type": "integer"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            },
            "description": "Widget updated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Validation error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Widget not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Update a widget"
      }
    }
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://localhost:3000"
    }
  ]
}